feat: 实现RBAC功能，支持角色继承

+++ api/service/internal/token/cliams.go
@@ -4,5 +4,5 @@ import `github.com/golang-jwt/jwt/v5`
 
 type UserToken struct {
     jwt.RegisteredClaims
-    UserID uint `json:"userID"`
+    UserID int64 `json:"userID"`
 }
